refactor: Hezarfen class created
